Refactor productController.js to create paymentController.js for better separation of concerns
