+++ routes/paymentRoutes.js
@@ -1,0 +1,18 @@
+import express from "express";
+import {
+  brainTreePaymentController,
+  braintreeTokenController,
+} from "../controllers/paymentController.js";
+import { requireSignIn } from "../middlewares/authMiddleware.js";
+
+const router = express.Router();
+
+//payments routes
+//token
+router.get("/braintree/token", braintreeTokenController);
+
+//payments
+router.post("/braintree/payment", requireSignIn, brainTreePaymentController);
+
+export default router;
+
